requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cssselect==1.1.0
fake-useragent==0.1.11
future==0.18.2
idna==2.8
lxml==4.4.2
parse==1.14.0
pycparser==2.19
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.1
PySocks==1.7.1
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
six==1.13.0
soupsieve==1.9.5
tornado==6.0.3
tqdm==4.41.1
urllib3==1.25.7
w3lib==1.21.0
websockets==8.1
